reparación completa

+++ firmware/Kboot/build/BUILD.sh
@@ -143,13 +143,11 @@ echo "}">> flash-list.json
 # Set a fixed timestamp for all files before zipping them and ensure reproducibility
 touch -t 200901031815 flash-list.json bootloader_lo.bin bootloader_hi.bin config.bin firmware.bin
 chmod 0644 flash-list.json bootloader_lo.bin bootloader_hi.bin config.bin firmware.bin
-zip kboot.kfpkg -9 flash-list.json bootloader_lo.bin bootloader_hi.bin config.bin firmware.bin > /dev/null
-rm -f flash-list.json
-
-if [ $? -ne 0 ]; then
-echo "ERROR creating kfpkg"
-exit 1
+if ! zip kboot.kfpkg -9 flash-list.json bootloader_lo.bin bootloader_hi.bin config.bin firmware.bin > /dev/null; then
+    echo "ERROR creating kfpkg"
+    exit 1
 fi
+rm -f flash-list.json
 
 
 # === CLEAN ===
